ethdb: moved DeleteRangeWithIterator logic to leveldb/memorydb
