+++ ethdb/leveldb/leveldb.go
@@ -21,6 +21,7 @@
 package leveldb
 
 import (
+	"bytes"
 	"fmt"
 	"sync"
 	"time"
@@ -206,28 +207,50 @@ func (db *Database) Delete(key []byte) error {
 	return db.db.Delete(key, nil)
 }
 
+var ErrTooManyKeys = errors.New("too many keys in deleted range")
+
 // DeleteRange deletes all of the keys (and values) in the range [start,end)
 // (inclusive on start, exclusive on end).
+// Note that this is a fallback implementation as leveldb does not natively
+// support range deletion. It can be slow and therefore the number of deleted
+// keys is limited in order to avoid blocking for a very long time.
+// ErrTooManyKeys is returned if the range has only been partially deleted.
+// In this case the caller can repeat the call until it finally succeeds.
 func (db *Database) DeleteRange(start, end []byte) error {
-	return ethdb.DeleteRangeWithIterator(db, start, end)
+	batch := db.NewBatch()
+	it := db.NewIterator(nil, start)
+	defer it.Release()
+
+	var count int
+	for it.Next() && bytes.Compare(end, it.Key()) > 0 {
+		count++
+		if count > 10000 { // should not block for more than a second
+			if err := batch.Write(); err != nil {
+				return err
+			}
+			return ErrTooManyKeys
+		}
+		if err := batch.Delete(it.Key()); err != nil {
+			return err
+		}
+	}
+	return batch.Write()
 }
 
 // NewBatch creates a write-only key-value store that buffers changes to its host
 // database until a final write is called.
 func (db *Database) NewBatch() ethdb.Batch {
 	return &batch{
-		db:        db.db,
-		wrappedDb: db,
-		b:         new(leveldb.Batch),
+		db: db.db,
+		b:  new(leveldb.Batch),
 	}
 }
 
 // NewBatchWithSize creates a write-only database batch with pre-allocated buffer.
 func (db *Database) NewBatchWithSize(size int) ethdb.Batch {
 	return &batch{
-		db:        db.db,
-		wrappedDb: db,
-		b:         leveldb.MakeBatch(size),
+		db: db.db,
+		b:  leveldb.MakeBatch(size),
 	}
 }
 
@@ -421,10 +444,9 @@ func (db *Database) meter(refresh time.Duration, namespace string) {
 // batch is a write-only leveldb batch that commits changes to its host database
 // when Write is called. A batch cannot be used concurrently.
 type batch struct {
-	db        *leveldb.DB
-	wrappedDb *Database
-	b         *leveldb.Batch
-	size      int
+	db   *leveldb.DB
+	b    *leveldb.Batch
+	size int
 }
 
 // Put inserts the given value into the batch for later committing.
